Resolve symlinks in project calculation
Some of Duncan's jobs failed, and we have narrowed it
down to missing projects in the job storage specification.

The projects were missed by tesp because his input paths
were syminks to other projects, and tesp was not doing the
mapping.

We need to add _both_ projects when a symlink straddles across two.